Replace `command -v` with `which`
If `foo` is a shell builtin, then `command -v foo` returns `foo` instead
of the full path to `foo`, which is undesirable when setting environment
variables. Therefore, we actually do want to shell out to `which` and
get the full path. (In Bash we could use `type -P foo`, but POSIX shells
like Dash aren't guaranteed to have that.)